refactor: Rank 패키지 이동 (utils)
