+++ src/main/java/lotto/domain/ticket/Ticket.java
@@ -1,5 +1,6 @@
 package lotto.domain.ticket;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
