+++ src/main/java/lotto/domain/ticket/Tickets.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import lotto.domain.ticket.generator.TicketGenerator;
-import lotto.domain.winning.Rank;
 import lotto.domain.winning.WinningTicket;
+import lotto.utils.Rank;
 
 public class Tickets {
