+++ src/main/java/lotto/view/output/OutputView.java
@@ -6,10 +6,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import lotto.domain.winning.Rank;
 import lotto.dto.AnalysisDto;
 import lotto.dto.TicketDto;
 import lotto.dto.TicketManagerDto;
+import lotto.utils.Rank;
 import lotto.view.utils.Delimiter;
 
 public class OutputView {
